move to typing.py
